refactor(zod): restructure package with flat src directory
Moves validators from nested validators/ and utils/ directories to src root
for improved import ergonomics and reduced cognitive overhead.

Structure changes:
- validators/* -> src/* (flattened hierarchy)
- utils/* -> src/* (consolidated with validators)
- Maintains all existing validator logic and exports
- Updates package.json exports configuration

Benefits:
- Shorter import paths: from '@atk/zod/validators/amount' to '@atk/zod/amount'
- Reduced directory nesting and navigation complexity
- Consistent with modern package conventions
- Maintains backward compatibility through package exports
