+++ packages/zod/src/equity-categories.ts
