+++ packages/zod/src/ethereum-hash.ts
