+++ packages/zod/src/fund-categories.ts
